mana-forge-engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
groq>=0.9.0
python-dotenv>=1.0.1
requests>=2.31.0